refactor routes to fix browser warning:
