+++ react/frontpage/Frontpage.jsx
@@ -51,27 +51,11 @@ Frontpage UI, including radio stream, navbar, and triangle background.
 Expects children components for tab contents
 
 @prop nowPlaying: show to display as live
-@prop spotlight: show to display as spotlight 
-@prop children: components to display in content area
+@prop spotlight: show to display as spotlight
 
-@prop updateNowPlaying: () => refresh now playing by requesting new data from server
-@prop setSpotlightShowID: (showID) => update the spotlight to be a show with the provided showID
-@prop updateShows: () => should fetch updated list of shows and update store
+@prop children: components to display in content area
 **/
 const FrontpageContent = React.createClass({
-  componentWillMount: function() {
-    // refresh live show info now and every 30 seconds
-    this.props.updateNowPlaying();
-    this.interval = setInterval(this.props.updateNowPlaying, 30*1000);
-
-    // update now playing and fetch initial shows data
-    // 1/11/17 - Black Twitter!
-    this.props.setSpotlightShowID(71);
-    this.props.updateShows();
-  },
-  componentWillUnmount: function() {
-    clearInterval(this.interval);
-  },
   render: function() {
     var showPlaying = this.props.nowPlaying && this.props.nowPlaying.title != null;
     return (
@@ -116,24 +100,80 @@ const FrontpageContent = React.createClass({
 });
 
 /**
-Wrapper for Frontpage content, implementing routes for different tabs.
-Will pass own props down to FrontpageContent
+Redux container (wrapper) for FrontpageContent to provide props:
+  nowPlaying, spotlight
+**/
+import { connect } from 'react-redux';
+const mapStateToProps = (state) => {
+  var props = {};
+
+  // set nowPlaying & spotlight shows if found
+  for (var showIndex = 0; showIndex < state.shows.shows.length; showIndex++) {
+    var show = state.shows.shows[showIndex];
+    if (show.id === state.shows.nowPlayingID) {
+      props.nowPlaying = show;
+    }
+    // note: only set spotlight show if different from current show
+    if (show.id === state.shows.spotlightID && show.id !== state.shows.nowPlayingID) {
+      props.spotlight = show;
+    }
+  }
+
+  var spotlight = state.shows.shows.find((show) => {
+    return show.id == state.shows.spotlightID;
+  });
+  if (spotlight && spotlight !== -1) {
+    props.spotlight = spotlight;
+  }
+
+  return props;
+};
+const FrontpageContentContainer = connect(
+  mapStateToProps
+)(FrontpageContent);
+
+// react-router routes 
+// each path represents a page of our frontpage with a component
+const routes = (
+  <Route path="/" component={FrontpageContentContainer}>
+    <IndexRoute component={WaterFallContent} />
+    <Route path="/djs" component={DJsTab} />
+    <Route path="/events" component={EventsTab} />
+    <Route path="/shows" components={ShowsTab} />
+    <Route path="/shows/:showID" component={ShowContainer} />
+    <Route path="/events/:eventID" component={EventContainer} />
+    <Route path="/streamIssues" component={StreamIssuesPage} />
+    <Route path="/about" component={AboutPage} />
+    <Route path="*" component={Error404Page} />
+  </Route>
+);
+
+/**
+Wrapper component, allows to use react-router to switch pages
+also sets some data in store
+
+@prop updateNowPlaying: () => refresh now playing by requesting new data from server
+@prop setSpotlightShowID: (showID) => update the spotlight to be a show with the provided showID
+@prop updateShows: () => should fetch updated list of shows and update store
 **/
 const Frontpage = React.createClass({
+  componentWillMount: function() {
+    // refresh live show info now and every 30 seconds
+    this.props.updateNowPlaying();
+    this.interval = setInterval(this.props.updateNowPlaying, 30*1000);
+
+    // update now playing and fetch initial shows data
+    // 1/11/17 - Black Twitter!
+    this.props.setSpotlightShowID(71);
+    this.props.updateShows();
+  },
+  componentWillUnmount: function() {
+    clearInterval(this.interval);
+  },
   render: function() {
     return (
       <Router history={browserHistory} onUpdate={logPageView}>
-        <Route path="/" component={props => <FrontpageContent {...this.props} {...props} />}>
-          <IndexRoute component={WaterFallContent} />
-          <Route path="/djs" component={DJsTab} />
-          <Route path="/events" component={EventsTab} />
-          <Route path="/shows" components={ShowsTab} />
-          <Route path="/shows/:showID" component={ShowContainer} />
-          <Route path="/events/:eventID" component={EventContainer} />
-          <Route path="/streamIssues" component={StreamIssuesPage} />
-          <Route path="/about" component={AboutPage} />
-          <Route path="*" component={Error404Page} />
-        </Route>
+        {routes}
       </Router>
     );
   }
